+++ examples/flax/run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+tango run config.jsonnet -d workspace --include-package xsum
